# src/main/resources/static/api-docs.yaml
openapi: 3.0.3
info:
  title: Phone numbers API v1
  description: Phone numbers API v1
  version: 1.0.0
servers:
  - url: 'http://localhost:8080/api/v1/'
paths:
  /phonenumbers:
    get:
      description: Returns phonenumbers with optional filtering and pagination
      operationId: searchPhoneNumbers
      parameters:
        - in: query
          name: customerName
          required: false
          schema:
            type: string
          example: Moe
        - in: query
          name: page
          required: false
          schema:
            type: integer
            default: 0
          example: 0
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 10
          example: 10
      responses:
        200:
          description: phonenumbers response
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalCount:
                    type: integer
                  records:
                    type: array
                    items:
                      $ref: '#components/schemas/PhoneNumber'
                example:
                  totalCount: 10
                  records:
                    - subscriberNumber: '455555551'
                      customerName: Moe
                      status: ACTIVE
                    - subscriberNumber: '455555552'
                      customerName: Moe
                      status: IN_ACTIVE


  /phonenumbers/{subscriberNumber}:
    patch:
      description: Updates an exisitng phonenumber
      operationId: updatePhoneNumber
      parameters:
        - in: path
          name: subscriberNumber
          required: true
          schema:
            type: string
          example: '455555551'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [ACTIVE,IN_ACTIVE]
              example:
                status: ACTIVE
      responses:
        200:
          description: Updated PhoneNumber resource
          content:
            application/json:
              schema:
                $ref: '#components/schemas/PhoneNumber'
              example:
                subscriberNumber: '455555551'
                customerName: Bill
                status: ACTIVE

components:
  schemas:
    PhoneNumber:
      type: object
      properties:
        subscriberNumber:
          type: string
        customerName:
          type: string
        status:
          type: string
          enum: [ACTIVE,IN_ACTIVE]